compression-example/requirements.txt
lz4
valkey-glide
numpy
python-dotenv

